backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.3.0
click==8.1.8
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.11
h11==0.14.0
idna==3.10
passlib==1.7.4
pyasn1==0.4.8
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.0
typing_extensions==4.12.2
uvicorn==0.34.0
